% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-genoscores.R
\name{.validate_input_genoscores}
\alias{.validate_input_genoscores}
\title{Validate user input and load GWAS — genoscores backend}
\usage{
.validate_input_genoscores(locus, src, trgt, gwas_type, refplinkfile, con)
}
\description{
Validate user input and load GWAS — genoscores backend
}
\keyword{internal}
