% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-genoscores.R
\name{find_missing_alleles_gs}
\alias{find_missing_alleles_gs}
\title{Resolve missing other alleles by querying the Genoscores dbSNP table}
\usage{
find_missing_alleles_gs(missing, con)
}
\arguments{
\item{missing}{A \code{data.table} of rows lacking \code{oth.allele}.}

\item{con}{DBI connection to the Genoscores database.}
}
\description{
Resolve missing other alleles by querying the Genoscores dbSNP table
}
\keyword{internal}
