% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-genoscores.R
\name{get_freqs_gs}
\alias{get_freqs_gs}
\title{Retrieve effect allele frequencies from the Genoscores r_freqs table}
\usage{
get_freqs_gs(gwas, con)
}
\arguments{
\item{gwas}{A \code{data.table} with columns \code{dbsnpid} and \code{allele}.}

\item{con}{DBI connection to the Genoscores database.}
}
\description{
Retrieve effect allele frequencies from the Genoscores r_freqs table
}
\keyword{internal}
