% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-genoscores.R
\name{get_published_gwas}
\alias{get_published_gwas}
\title{Read published GWAS flatfiles from the Genoscores datastore}
\usage{
get_published_gwas(gwasids, con)
}
\arguments{
\item{gwasids}{Integer vector of gwasids to retrieve.}

\item{con}{DBI connection to the Genoscores database.}
}
\description{
Read published GWAS flatfiles from the Genoscores datastore
}
\keyword{internal}
