% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-genoscores.R
\name{liftoverbim}
\alias{liftoverbim}
\title{Lift genomic positions to hg38 using liftOver}
\usage{
liftoverbim(bim, liftover_file)
}
\arguments{
\item{bim}{A \code{data.table} with columns \code{snp}, \code{chrom}, \code{pos}.}

\item{liftover_file}{Full path to the liftOver chain file.}
}
\description{
Lift genomic positions to hg38 using liftOver
}
\keyword{internal}
