% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-genoscores.R
\name{process_private_gwas}
\alias{process_private_gwas}
\title{Post-process a raw SNPTEST GWAS}
\usage{
process_private_gwas(gwas, outcome)
}
\description{
Post-process a raw SNPTEST GWAS
}
\keyword{internal}
